core: relocate `layout` to `components`
