+++ packages/admin/app.tsx
@@ -1,20 +1,27 @@
 import { AppProvider } from './contexts/AppProvider';
+import {
+	ErrorBoundary,
+	Header,
+	Screen,
+	ScreenProvider,
+} from './components/layout';
 import { Notices } from './components/notices';
-import { Header, Screen, ScreenProvider } from './layout';
 
 import './admin.scss';
 
 const App = () => {
 	return (
-		<AppProvider>
-			<ScreenProvider>
-				<Header />
-				<Screen />
-			</ScreenProvider>
-			<div className="wp-graphql-headless-login__notices">
-				<Notices />
-			</div>
-		</AppProvider>
+		<ErrorBoundary showErrorInfo>
+			<AppProvider>
+				<ScreenProvider>
+					<Header />
+					<Screen />
+				</ScreenProvider>
+				<div className="wp-graphql-headless-login__notices">
+					<Notices />
+				</div>
+			</AppProvider>
+		</ErrorBoundary>
 	);
 };
