+++ packages/admin/components/layout/index.ts
@@ -1,3 +1,4 @@
+export * from './error-boundary';
 export * from './header';
 export * from './screen';
 export * from './screen/context';
