+++ packages/admin/index.tsx
@@ -4,7 +4,6 @@ import { createHooks } from '@wordpress/hooks';
 import { StrictMode } from 'react';
 
 import App from './app';
-import { ErrorBoundary } from './layout/error-boundry/error-boundry';
 
 export const hooks = createHooks();
 
@@ -20,9 +19,7 @@ domReady( () => {
 
 	root.render(
 		<StrictMode>
-			<ErrorBoundary showErrorInfo>
-				<App />
-			</ErrorBoundary>
+			<App />
 		</StrictMode>
 	);
 } );
